+++ user/src/main/java/daehee/challengehub/authentication/model/UserSignupDto.java
@@ -1,4 +1,4 @@
-package daehee.challengehub.model;
+package daehee.challengehub.authentication.model;
 
 import lombok.Builder;
 import lombok.Getter;
